add latest cmake to version requirements
this prevents current versions of cmake from erroring out due to out-of-date version requirements. we could have just changed the minimum to 3.5, but that results in a further deprecation warning; instead, adding the max gives the most amount of time until someone needs to update this again, and as a bonus should remain compatible with older cmake versions. in theory at least, we're not a cmake guru